requirements-base.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
cron-descriptor==1.4.5
cymem==2.0.11
Cython==3.0.12
cytoolz==1.0.1
Django==5.1.7
django-appconf==1.1.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-timezone-field==7.1
django-tus==0.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
eth-hash==0.7.1
eth-keys==0.6.1
eth-typing==5.2.0
eth-utils==5.2.0
filelock==3.17.0
flower==2.0.1
fsspec==2025.2.0
hiredis==3.1.0
honcho==2.0.0
huggingface-hub==0.29.2
humanfriendly==10.0
humanize==4.12.1
idna==3.10
inflection==0.5.1
joblib==1.4.2
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
ml-dtypes==0.4.1
more-itertools==10.6.0
murmurhash==1.0.12
packaging==24.2
pathvalidate==2.3.0
pillow==11.1.0
ply==3.11
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-crontab==3.2.0
pdfplumber
python-docx
python-dateutil==2.9.0.post0
python-ulid==3.0.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
spacy==3.8.5
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
types-cffi==1.16.0.20250307
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==75.8.2.20250305
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
walrus==0.9.4
wcwidth==0.2.13
